+++ zebra-state/src/service.rs
@@ -56,8 +56,9 @@ use crate::{
         read::find,
         watch_receiver::WatchReceiver,
     },
-    BoxError, CheckpointVerifiedBlock, CommitSemanticallyVerifiedError, Config, KnownBlock,
-    ReadRequest, ReadResponse, Request, Response, SemanticallyVerifiedBlock, StateInitError,
+    AwaitUtxoError, BoxError, CheckpointVerifiedBlock, CommitSemanticallyVerifiedError, Config,
+    KnownBlock, ReadRequest, ReadResponse, Request, Response, SemanticallyVerifiedBlock,
+    StateInitError,
 };
 
 pub mod block_iter;
@@ -1163,14 +1164,19 @@ impl Service<Request> for StateService {
 
             // Uses pending_utxos and non_finalized_state_queued_blocks in the StateService.
             // If the UTXO isn't in the queued blocks, runs concurrently using the ReadStateService.
+            //
+            // The expected error type for this request is `AwaitUtxoError`.
             Request::AwaitUtxo(outpoint) => {
                 let timer = CodeTimer::start();
-                // Prepare the AwaitUtxo future from PendingUxtos.
+                // Prepare the AwaitUtxo future from PendingUtxos.
                 let response_fut = self.pending_utxos.queue(outpoint);
                 // Only instrument `response_fut`, the ReadStateService already
                 // instruments its requests with the same span.
 
-                let response_fut = response_fut.instrument(span).boxed();
+                let response_fut = response_fut
+                    .map(|result| result.map_err(BoxError::from))
+                    .instrument(span)
+                    .boxed();
 
                 // Check the non-finalized block queue outside the returned future,
                 // so we can access mutable state fields.
@@ -1209,7 +1215,10 @@ impl Service<Request> for StateService {
                 async move {
                     let req = ReadRequest::AnyChainUtxo(outpoint);
 
-                    let rsp = read_service.oneshot(req).await?;
+                    let rsp = read_service
+                        .oneshot(req)
+                        .await
+                        .map_err(AwaitUtxoError::ReadStateFailed)?;
 
                     // Optional TODO:
                     //  - make pending_utxos.respond() async using a channel,