+++ zebra-state/CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+All notable changes to this project are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+and this project adheres to [Semantic Versioning](https://semver.org).
 
-## [Unreleased]
+### Added
+
+- Added the concrete error type `AwaitUtxoError` for handling failed `AwaitUtxo` state requests
+  [#11205](https://github.com/ZcashFoundation/zebra/pull/11205).
+- Added `MappedRequest` for `AwaitUtxoRequest`
+  [#11205](https://github.com/ZcashFoundation/zebra/pull/11205).
 
 ## [13.0.0] - 2026-08-10